+++ packages/contentful--app-scripts/src/upload/build-upload-settings.ts
@@ -37,6 +37,7 @@ export async function buildAppUploadSettings(options: UploadOptions): Promise<Up
       name: 'host',
       message: `Contentful CMA endpoint URL:`,
       default: DEFAULT_CONTENTFUL_API_HOST,
+      filter: hostProtocolFilter,
     });
   }
 
@@ -54,4 +55,8 @@ export async function buildAppUploadSettings(options: UploadOptions): Promise<Up
     ...appUploadSettings,
     ...appInfo,
   };
+}
+
+export function hostProtocolFilter (input: string) {
+  return input.replace(/^https?:\/\//, '');
 }